Restructured dirs.
