+++ gits/entities/aligned/aligned_image.py
