--- gits/entities/aligned_roi.py
@@ -1,1 +1,0 @@
-#!/usr/bin/env python
