+++ gits/entities/interfaces/roi_interface.py
