+++ gits/entities/interfaces/scene_interface.py
