+++ gits/entities/scene_id.py
@@ -1,4 +1,3 @@
-
 import calendar
 
 from utils import logging
